SAM_GUI.mlapp
Options Tab
Output Variables: max energy doesn't work when included in outputNames (Line 192), but is added back in for plotting purpuses in line 209.  Any desired output variables that don't need to be specified in MAIN can be added to line 209 as well.
Output File Type drop down: currently disabled.  Plan for CSV file types (xlsread/write are more inefficient), but we still need to change several places in the code for this to be possible.
Compile Results button: runs Master_Driver code, with additions in Data_Format_Interpolation and Output_Driver (second part of Master_Driver where output files are generated) to check if any of the files have already been compiled, since it takes so long to run.

Advanced Tab
Saturation Current Temperature Dependency Constants: currently disabled, but variables are stored and sent to relevant functions in the code.
External Quantum Efficiency: enabled.  User can choose a file if custom values are desired (not tested), but the format of the file needs to match that in Sub_Surface_Solar_Power (Line 16).

Output Tab
User can load a water quality file and then select variables to plot, time range, and plot type.  Once the plot is generated (Plot Variables), the user can open the figure, giving them options to edit or save as any file type, or just save as a PDF with one click.

Current Issues
Running out of memory in MATLAB or crashing the computer when all variables are selected.
I tried converting many variables to single precision, but the output gave complex numbers for many variables, so I converted some back to double precision.